requirements.txt
python-telegram-bot==20.6
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
certifi==2023.11.17
APScheduler==3.10.4 